receiveImage/requirements.txt
cachetools==5.3.0
docopt==0.6.2
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-cloud-tasks==2.12.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
mkl-fft==1.3.1
mkl-service==2.4.0
pandas==1.3.5
pipreqs==0.4.11
proto-plus==1.22.2
protobuf==4.21.12
